add bevel & remesh option.

+++ 3DTextMaker/test.py
@@ -8,22 +8,41 @@ dirname = "C:/.../output"
 fontpath = "C:/Windows/Fonts/meiryo.ttc"
 text = "3DTextMaker試作@DMiyamo3"
 
-def test(i, text, extrude = 0.05, bevel_depth = 0.05, offset = 0, bevel_resolution = 4):
+def test(i, text, isSave = True, isBevel = False, isRemesh = False):
     bpy.ops.object.text_add(view_align=False, enter_editmode=False, location=(0, 0, 0))
-    bpy.context.object.data.extrude = extrude
-    bpy.context.object.data.bevel_depth = bevel_depth
-    bpy.context.object.data.offset = offset
-    bpy.context.object.data.bevel_resolution = bevel_resolution
-
     txt = bpy.data.objects['Text']
     txt.data.body = text
     fnt = bpy.data.fonts.load(fontpath)
     txt.data.font = fnt
 
-    filename = os.path.join(dirname,str(i) + "_" + text + ".fbx")
-    bpy.ops.export_scene.fbx(filepath=filename, use_selection=True)
+    # extrude
+    bpy.context.object.data.extrude = 0.05
 
-    bpy.ops.object.delete()
-    
+    # offset
+    bpy.context.object.data.offset = 0
+
+    # bevel
+    if isBevel:
+        bpy.context.object.data.bevel_depth = 0.05
+        bpy.context.object.data.bevel_resolution = 4
+
+    # remesh
+    if isRemesh:
+        bpy.ops.object.modifier_add(type='REMESH')
+        bpy.context.object.modifiers["Remesh"].mode = 'BLOCKS'
+        bpy.context.object.modifiers["Remesh"].octree_depth = 4
+        bpy.context.object.modifiers["Remesh"].scale = 0.99
+        bpy.context.object.modifiers["Remesh"].use_remove_disconnected = False
+
+    # save fbx
+    if isSave:
+        filename = os.path.join(dirname,str(i) + "_" + text + ".fbx")
+        bpy.ops.export_scene.fbx(filepath=filename, use_selection=True)
+        bpy.ops.object.delete()
+
+# main
 for i,c in enumerate(text):
     test(i, c)
+
+# test 確認用
+# test(text)
